fix(fiori-gen): move writing to fiori gen shared
