+++ packages/fiori-app-sub-generator/test/int/fiori-elements/expected-output/lrop_v2_dest/.appGenInfo.json
@@ -1,6 +1,6 @@
 {
   "generationParameters": {
-    "generationDate": "Fri Jun 06 2025 13:53:11 GMT+0100 (Irish Standard Time)",
+    "generationDate": "Fri Jun 06 2025 15:22:17 GMT+0100 (Irish Standard Time)",
     "generatorPlatform": "CLI",
     "serviceType": "MTA File",
     "metadataFilename": "",
@@ -33,7 +33,7 @@
         "packageUri": "abapfs:/BAS_DEST/MOCK_CSN_TEST1",
         "csnName": "MOCKCSN55.abap.csn",
         "serviceNameCsn": "MockService2",
-        "serviceUri": "/sap/opu/odata/sap/SEPMRA_PROD_MAN"
+        "serviceUrl": "https://sap-ux-mock-services-v2-lrop.cfapps.us10.hana.ondemand.com/sap/opu/odata/sap/SEPMRA_PROD_MAN"
       }
     ]
   }
